About Argo
Argo is a recycling company located in Roskilde. 
This project, is an android application for Argo. The main goal of the application is to help users know about the different plastic types, and learn more about each type.
The project uses Firebase Realtime database, so the user can search for specific item and know its type. 
Languages used in this app are Java, XML.
The project is based on Single Activity Architecture, and Navigation